Filters optimization

+++ Firmware/FFBoard/Src/Filters.cpp
@@ -54,97 +54,97 @@ void Biquad::calcBiquad(void) {
 	z1 = 0.0;
 	z2 = 0.0;
     float norm;
-    float V = pow(10, fabs(peakGain) / 20.0);
-    float K = tan(M_PI * Fc);
+    float V = powf(10, fabsf(peakGain) / 20.0f);
+    float K = tanf((float)M_PI * Fc);
     switch (this->type) {
         case BiquadType::lowpass:
-            norm = 1 / (1 + K / Q + K * K);
+            norm = 1.0f / (1.0f + K / Q + K * K);
             a0 = K * K * norm;
-            a1 = 2 * a0;
+            a1 = 2.0f * a0;
             a2 = a0;
-            b1 = 2 * (K * K - 1) * norm;
-            b2 = (1 - K / Q + K * K) * norm;
+            b1 = 2.0f * (K * K - 1.0f) * norm;
+            b2 = (1.0f - K / Q + K * K) * norm;
             break;
 
         case BiquadType::highpass:
-            norm = 1 / (1 + K / Q + K * K);
-            a0 = 1 * norm;
-            a1 = -2 * a0;
+            norm = 1.0f / (1.0f + K / Q + K * K);
+            a0 = 1.0f * norm;
+            a1 = -2.0f * a0;
             a2 = a0;
-            b1 = 2 * (K * K - 1) * norm;
-            b2 = (1 - K / Q + K * K) * norm;
+            b1 = 2.0f * (K * K - 1.0f) * norm;
+            b2 = (1.0f - K / Q + K * K) * norm;
             break;
 
         case BiquadType::bandpass:
-            norm = 1 / (1 + K / Q + K * K);
+            norm = 1.0f / (1.0f + K / Q + K * K);
             a0 = K / Q * norm;
             a1 = 0;
             a2 = -a0;
-            b1 = 2 * (K * K - 1) * norm;
-            b2 = (1 - K / Q + K * K) * norm;
+            b1 = 2.0f * (K * K - 1.0f) * norm;
+            b2 = (1.0f - K / Q + K * K) * norm;
             break;
 
         case BiquadType::notch:
-            norm = 1 / (1 + K / Q + K * K);
-            a0 = (1 + K * K) * norm;
-            a1 = 2 * (K * K - 1) * norm;
+            norm = 1.0f / (1.0f + K / Q + K * K);
+            a0 = (1.0f + K * K) * norm;
+            a1 = 2.0f * (K * K - 1.0f) * norm;
             a2 = a0;
             b1 = a1;
-            b2 = (1 - K / Q + K * K) * norm;
+            b2 = (1.0f - K / Q + K * K) * norm;
             break;
 
         case BiquadType::peak:
             if (peakGain >= 0) {    // boost
-                norm = 1 / (1 + 1/Q * K + K * K);
-                a0 = (1 + V/Q * K + K * K) * norm;
-                a1 = 2 * (K * K - 1) * norm;
-                a2 = (1 - V/Q * K + K * K) * norm;
+                norm = 1.0f / (1.0f + 1.0f/Q * K + K * K);
+                a0 = (1.0f + V/Q * K + K * K) * norm;
+                a1 = 2.0f * (K * K - 1) * norm;
+                a2 = (1.0f - V/Q * K + K * K) * norm;
                 b1 = a1;
-                b2 = (1 - 1/Q * K + K * K) * norm;
+                b2 = (1.0f - 1.0f/Q * K + K * K) * norm;
             }
             else {    // cut
-                norm = 1 / (1 + V/Q * K + K * K);
-                a0 = (1 + 1/Q * K + K * K) * norm;
-                a1 = 2 * (K * K - 1) * norm;
-                a2 = (1 - 1/Q * K + K * K) * norm;
+                norm = 1.0f / (1.0f + V/Q * K + K * K);
+                a0 = (1.0f + 1.0f/Q * K + K * K) * norm;
+                a1 = 2.0f * (K * K - 1.0f) * norm;
+                a2 = (1.0f - 1.0f/Q * K + K * K) * norm;
                 b1 = a1;
-                b2 = (1 - V/Q * K + K * K) * norm;
+                b2 = (1.0f - V/Q * K + K * K) * norm;
             }
             break;
         case BiquadType::lowshelf:
             if (peakGain >= 0) {    // boost
-                norm = 1 / (1 + sqrt(2) * K + K * K);
-                a0 = (1 + sqrt(2*V) * K + V * K * K) * norm;
-                a1 = 2 * (V * K * K - 1) * norm;
-                a2 = (1 - sqrt(2*V) * K + V * K * K) * norm;
-                b1 = 2 * (K * K - 1) * norm;
-                b2 = (1 - sqrt(2) * K + K * K) * norm;
+                norm = 1.0f / (1.0f + sqrtf(2.0f) * K + K * K);
+                a0 = (1.0f + sqrtf(2.0f*V) * K + V * K * K) * norm;
+                a1 = 2.0f * (V * K * K - 1.0f) * norm;
+                a2 = (1.0f - sqrtf(2.0f*V) * K + V * K * K) * norm;
+                b1 = 2.0f * (K * K - 1.0f) * norm;
+                b2 = (1.0f - sqrtf(2.0f) * K + K * K) * norm;
             }
             else {    // cut
-                norm = 1 / (1 + sqrt(2*V) * K + V * K * K);
-                a0 = (1 + sqrt(2) * K + K * K) * norm;
-                a1 = 2 * (K * K - 1) * norm;
-                a2 = (1 - sqrt(2) * K + K * K) * norm;
-                b1 = 2 * (V * K * K - 1) * norm;
-                b2 = (1 - sqrt(2*V) * K + V * K * K) * norm;
+                norm = 1.0f / (1.0f + sqrtf(2.0f*V) * K + V * K * K);
+                a0 = (1.0f + sqrtf(2.0f) * K + K * K) * norm;
+                a1 = 2.0f * (K * K - 1.0f) * norm;
+                a2 = (1.0f - sqrtf(2.0f) * K + K * K) * norm;
+                b1 = 2.0f * (V * K * K - 1) * norm;
+                b2 = (1.0f - sqrtf(2.0f*V) * K + V * K * K) * norm;
             }
             break;
         case BiquadType::highshelf:
             if (peakGain >= 0) {    // boost
-                norm = 1 / (1 + sqrt(2) * K + K * K);
-                a0 = (V + sqrt(2*V) * K + K * K) * norm;
-                a1 = 2 * (K * K - V) * norm;
-                a2 = (V - sqrt(2*V) * K + K * K) * norm;
-                b1 = 2 * (K * K - 1) * norm;
-                b2 = (1 - sqrt(2) * K + K * K) * norm;
+                norm = 1.0f / (1.0f + sqrtf(2.0f) * K + K * K);
+                a0 = (V + sqrtf(2.0f*V) * K + K * K) * norm;
+                a1 = 2.0f * (K * K - V) * norm;
+                a2 = (V - sqrtf(2.0f*V) * K + K * K) * norm;
+                b1 = 2.0f * (K * K - 1.0f) * norm;
+                b2 = (1.0f - sqrtf(2.0f) * K + K * K) * norm;
             }
             else {    // cut
-                norm = 1 / (V + sqrt(2*V) * K + K * K);
-                a0 = (1 + sqrt(2) * K + K * K) * norm;
-                a1 = 2 * (K * K - 1) * norm;
-                a2 = (1 - sqrt(2) * K + K * K) * norm;
-                b1 = 2 * (K * K - V) * norm;
-                b2 = (V - sqrt(2*V) * K + K * K) * norm;
+                norm = 1.0f / (V + sqrtf(2.0f*V) * K + K * K);
+                a0 = (1.0f + sqrtf(2.0f) * K + K * K) * norm;
+                a1 = 2.0f * (K * K - 1.0f) * norm;
+                a2 = (1.0f - sqrtf(2.0f) * K + K * K) * norm;
+                b1 = 2.0f * (K * K - V) * norm;
+                b2 = (V - sqrtf(2.0f*V) * K + K * K) * norm;
             }
             break;
     }
